clearer structrue
